Generalize heuristic for using display names in XML reports
This change partially reverts #13250, in particular the changes done to
`JUnitPlatformTestExecutionListener` which tried to determine whether a
custom display name was used based on the `DisplayName` annotation from
JUnit Jupiter. There were a couple of problems with that approach: it
did not support other test engines which might have other annotations,
it did not support composed annotations or other ways of setting the
display name (e.g. Jupiter's `DisplayNameGenerator` API).

We now use a simple heuristic that checks whether the class display name
is the simple name of the test class which is used as the default
display name for Jupiter, Vintage, and other test engines. Instead of
applying this heuristic eagerly and disposing the display name, we now
do it in `JUnitXmlResultWriter` which allows to reuse it for the
test-distribution plugin.